Added BottomNavigation Bar
Reorganized the file structure and added riverpod dependencies.
