+++ toura/lib/main.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:toura/src/toura.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:toura/UI/toura.dart';
 
 void main() {
-  runApp(const Toura());
+  runApp(const ProviderScope(child: Toura()));
 }
